refactor pdf to html functions
