+++ .gitignore
@@ -169,3 +169,4 @@ cython_debug/
 .temp
 logs
 *.sqlite
+output*
