+++ CHANGELOG.md
@@ -3,3 +3,19 @@
 ## 2023-08-24
 
 - setup project
+- create function to convert pdf to docx
+- create function to convert docx to html
+
+## 2023-08-25
+
+- update `pdfweb.__init__` to include `pdfweb.pdftoweb` functions
+
+## 2023-08-27
+
+- created `pdfweb.defaults`
+- add license
+
+## 2023-08-28
+
+- refactor function to convert pdf to docx
+- refactor function to convert docx to html
